+++ Source/Forms/Util/EncounterLookup.hpp
@@ -45,8 +45,8 @@ private:
     QStandardItemModel *model = nullptr;
 
     void setupModels();
-    std::set<std::pair<u8, QString>> getEncounters3(Game game, u16 specie);
-    std::set<std::pair<u8, QString>> getEncounters4(Game game, u16 specie);
+    std::set<std::pair<u16, QString>> getEncounters3(Game game, u16 specie);
+    std::set<std::pair<u16, QString>> getEncounters4(Game game, u16 specie);
     QString getEncounterString(Encounter type);
 
 private slots:
